TokenVault: Improving the emergency token recover mechanism
Nowadays we have the "Recoverable" mechanism, which is designed to recover
unintended tokens from a contract. TokenVault also implements this, but it's
only partial: it is designed to recover extra tokens of the token for which
the vault was created. This commit is about expanding tokensToBeReturned() to
support all the tokens.